added a comment for after method sleep and moved LogListener
